requirements.txt
aws-cdk-lib~=2.77
pyyaml~=6.0.1
boto3
